pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "ghosttrace"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
	"click",
	"httpx",
	"dnspython",
	"python-whois",
	]

[project.scripts]
ghosttrace = "ghosttrace.cli:main"